+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.4.1 / YYYY-MM-DD
 
+* Bump `commons-digester` from 1.8 to 3.2
 * Bump `webdrivermanager` from 5.2.1 to 5.2.3
 * Bump `htmlunit` from 2.63.0 to 2.64.0
 * Bump `commons-codec` from 1.2 to 1.15